Split MonadTime out of MonadTimer
And make them independent classes, removing the subclass/superclass
relationship.
