+++ io-sim-classes/io-sim-classes.cabal
@@ -28,6 +28,7 @@ library
                        Control.Monad.Class.MonadST
                        Control.Monad.Class.MonadSTM
                        Control.Monad.Class.MonadThrow
+                       Control.Monad.Class.MonadTime
                        Control.Monad.Class.MonadTimer
   default-language:    Haskell2010
   other-extensions:    CPP
